gemma3-fine-tuning-and-inference/requirements.txt
numpy==1.26.4
torch>=2.4.0
tensorboard
torchvision
git+https://github.com/huggingface/transformers@v4.49.0-Gemma-3
datasets==3.3.2
accelerate==1.4.0
evaluate==0.4.3
bitsandbytes==0.45.3
trl==0.15.2
peft==0.14.0
pillow==11.1.0
wandb==0.15.12
protobuf
sentencepiece
